requirements.txt
gradio>=4.29.0
openai>=1.30.0
azure-ai-textanalytics>=5.3.0
azure-cognitiveservices-vision-computervision>=0.9.0
azure-search-documents>=11.4.0
azure-ai-translation-text>=1.0.0
azure-ai-contentsafety>=1.0.0
azure-identity>=1.16.0
azure-keyvault-secrets>=4.8.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.3.0
httpx>=0.27.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
faiss-cpu>=1.8.0
tiktoken>=0.7.0
pydantic>=2.7.0
rich>=13.7.0
